Add comment update and verification for delete

+++ controllers/postController.js
@@ -1,7 +1,6 @@
 const Post = require("../models/Post");
 const Comment = require("../models/Comment");
 const asyncHandler = require("express-async-handler");
-const mongoose = require("mongoose");
 const { body, validationResult } = require("express-validator");
 const jwt = require("jsonwebtoken");
 require("dotenv").config();
@@ -13,6 +12,7 @@ exports.list = asyncHandler(async (req, res) => {
   }
   res.json(posts);
 });
+
 exports.detail = asyncHandler(async (req, res) => {
   const post = await Post.findById(req.params.postid).exec();
   if (post === null) {
@@ -20,6 +20,7 @@ exports.detail = asyncHandler(async (req, res) => {
   }
   res.json(post);
 });
+
 exports.create = [
   body("title", "title must not be empty").isLength({ min: 1 }).escape(),
   body("text", "text must be longer than 5 characters")
@@ -28,11 +29,15 @@ exports.create = [
   asyncHandler(async (req, res) => {
     const result = validationResult(req);
     if (!result.isEmpty()) {
-      return res.json({ error: result.array()[0].msg });
+      return res.status(400).json({
+        error: "Validation failed",
+        message: result.array()[0].msg,
+      });
     }
     jwt.verify(req.token, process.env.secretkey, async (err, authData) => {
+      console.log(authData.user);
       if (err) {
-        res.sendStatus(403);
+        return res.status(403).json({ error: "You must be logged in" });
       } else {
         const exists = await Post.findOne({ title: req.body.title });
         if (exists !== null)
@@ -41,7 +46,7 @@ exports.create = [
           title: req.body.title,
           text: req.body.text,
           author: authData.user._id,
-          published: false,
+          published: req.body.published,
         });
         await post.save();
         res.json(post);
@@ -49,11 +54,56 @@ exports.create = [
     });
   }),
 ];
+
+exports.update = [
+  body("title", "title must not be empty").isLength({ min: 1 }).escape(),
+  body("text", "text must be longer than 5 characters")
+    .isLength({ min: 5 })
+    .escape(),
+  asyncHandler(async (req, res) => {
+    const result = validationResult(req);
+    if (!result.isEmpty())
+      return res
+        .status(400)
+        .json({ error: "Validation failed", message: result.array()[0].msg });
+    if (post === null)
+      return res.status(400).json({ error: "Post doesn't exist" });
+    jwt.verify(req.token, process.env.secretkey, async (err, authData) => {
+      if (err) {
+        return res.status(403).json({ error: "You must be logged in" });
+      }
+      const post = await Post.findById(req.params.postid);
+      if (post.author._id.toString() !== authData.user._id)
+        return res.status(403).json({ error: "You're not an author" });
+      const editedPost = {
+        title: req.body.title,
+        text: req.body.text,
+        published: req.body.published,
+      };
+      const updated = await Post.findOneAndUpdate(
+        { _id: req.params.postid },
+        editedPost,
+        {
+          new: true,
+        }
+      );
+      res.json({ message: "Succesfully updated!", updated });
+    });
+  }),
+];
+
 exports.delete = asyncHandler(async (req, res) => {
   const post = await Post.findById(req.params.postid);
   if (post === null)
     return res.status(400).json({ error: "Post doesn't exist" });
-  await Post.findByIdAndDelete(req.params.postid);
-  await Comment.deleteMany({ post: req.params.postid });
-  res.json("Post deleted");
+  jwt.verify(req.token, process.env.secretkey, async (err, authData) => {
+    if (err) {
+      return res.status(403).json({ error: "You must be logged in" });
+    }
+    if (post.author._id.toString() !== authData.user._id)
+      return res.status(403).json({ error: "You're not an author" });
+    await Post.findByIdAndDelete(req.params.postid);
+    await Comment.deleteMany({ post: req.params.postid });
+    res.json("Post deleted");
+  });
 });
